re-adding unlink

+++ bin/server/cognizen-socket-handler.js
@@ -85,7 +85,7 @@ var SocketHandler = {
                                         });
 
                                         stream.on('close', function(){
-                                            fs.unlink(event.file.pathName);
+                                            if (!had_error) fs.unlink(event.file.pathName);
                                         })
                                         //Git commit
                                     } else {
